requirements.txt
# Core Framework
langchain>=0.1.0
langgraph>=0.0.40
langchain-google-genai>=1.0.0

# Google AI
# Align with langchain-google-genai enums (GenerationConfig.MediaResolution)
google-generativeai>=0.8.3

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0
streamlit>=1.31.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.0
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
aiosqlite>=0.19.0  # For SQLite async support

# Caching & Queue
redis>=5.0.1
celery>=5.3.4

# Image Processing
pillow>=10.2.0
opencv-python>=4.9.0
scikit-image>=0.22.0

# Cloud Storage
google-cloud-storage>=2.14.0

# HTTP Clients
httpx>=0.26.0
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Monitoring & Logging
structlog>=24.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0  # For testing FastAPI

# Development
black>=24.1.0
ruff>=0.1.0
mypy>=1.8.0
pre-commit>=3.6.0

# Payment Processing (for future)
stripe>=7.0.0

# Vector Database (for future semantic search)
pgvector>=0.2.4
sentence-transformers>=3.0.1

